% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_report_numbers.R
\name{add_report_numbers}
\alias{add_report_numbers}
\title{Add the report numbers to a bookdown report}
\usage{
add_report_numbers(
  path,
  doi,
  year,
  reportnr,
  depotnr,
  embargo = Sys.Date(),
  ordernr,
  copies,
  motivation,
  pages
)
}
\arguments{
\item{path}{the path to the \code{index.Rmd} of the bookdown report.}

\item{doi}{a string containing the DOI.}

\item{year}{Publication year}

\item{reportnr}{Report number assigned by the INBO library.}

\item{depotnr}{Report number assigned by the INBO library.}

\item{embargo}{The earliest date at which the report can be made public on
the INBO website.
Defaults to today.}

\item{ordernr}{Optional reference number specified by the client.}

\item{copies}{Set the required number of copies in case you really need a
printed version of the report.
Omit in case you only require a digital version.
Keep in mind that the INBO policy is to be "radically digital".
Which implies to only print a report if it is mandatory.}

\item{motivation}{Required motivation in case \code{copies} is set to a non-zero
number.}

\item{pages}{Number of pages of the report.
Only required in case \code{copies} is set to a non-zero number.}
}
\description{
Add the report numbers to a bookdown report
}
\seealso{
Other utils: 
\code{\link{check_dependencies}()},
\code{\link{column_start}()},
\code{\link{column_width}()},
\code{\link{cover_info}()},
\code{\link{create_report}()},
\code{\link{dyn_table}()},
\code{\link{references}()},
\code{\link{render_natbib}()},
\code{\link{validate_doi}()}
}
\concept{utils}
