% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insert_block}
\alias{insert_block}
\alias{insert_exampleblock}
\alias{insert_alertblock}
\title{Insert a markdown block
These functions will add template chunks into your \code{.Rmd} file.
A markdown block will be display as a frame with a title in a different
background.
There are 3 types of blocks: \code{"block"}, \code{"exampleblock"} and \code{"alertblock"}.}
\usage{
insert_block()

insert_exampleblock()

insert_alertblock()
}
\description{
Insert a markdown block
These functions will add template chunks into your \code{.Rmd} file.
A markdown block will be display as a frame with a title in a different
background.
There are 3 types of blocks: \code{"block"}, \code{"exampleblock"} and \code{"alertblock"}.
}
\concept{addins}
