% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_width.R
\name{column_start}
\alias{column_start}
\title{Calculate the starting position of an A0 column}
\usage{
column_start(number = 3, start = 1, margin = 20, edge = 10)
}
\arguments{
\item{number}{The desired number of columns with equal width.}

\item{start}{the number of the column}

\item{margin}{the horizontal distance between columns in mm.}

\item{edge}{the width of the whitespace edge around the poster in mm.}
}
\description{
Calculate the starting position of an A0 column
}
\seealso{
Other utils: 
\code{\link{add_author}()},
\code{\link{add_report_numbers}()},
\code{\link{check_dependencies}()},
\code{\link{column_width}()},
\code{\link{cover_info}()},
\code{\link{create_report}()},
\code{\link{dyn_table}()},
\code{\link{references}()},
\code{\link{render_natbib}()},
\code{\link{validate_doi}()}
}
\concept{utils}
