% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_width.R
\name{column_width}
\alias{column_width}
\title{Calculate the width of an A0 column}
\usage{
column_width(number = 3, margin = 20, edge = 10)
}
\arguments{
\item{number}{The desired number of columns with equal width.}

\item{margin}{the horizontal distance between columns in mm.}

\item{edge}{the width of the whitespace edge around the poster in mm.}
}
\description{
Calculate the width of an A0 column
}
\seealso{
Other utils: 
\code{\link{check_dependencies}()},
\code{\link{column_start}()},
\code{\link{cover_info}()},
\code{\link{create_report}()},
\code{\link{dyn_table}()},
\code{\link{references}()},
\code{\link{render_natbib}()},
\code{\link{validate_doi}()}
}
\concept{utils}
