% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_info.R
\name{cover_info}
\alias{cover_info}
\title{Create a cover.txt from the YAML header}
\usage{
cover_info(path = "index.Rmd")
}
\arguments{
\item{path}{The path to the main \code{.Rmd} file.}
}
\description{
You need to provide information to create the cover image and (optionally)
print the report.
Add the information to the YAML header of the report.
Then run this function to covert the information into a \code{cover.txt} and
send that file along the other files to the review process.
}
\seealso{
Other utils:
\code{\link[=add_author]{add_author()}},
\code{\link[=add_report_numbers]{add_report_numbers()}},
\code{\link[=check_dependencies]{check_dependencies()}},
\code{\link[=column_start]{column_start()}},
\code{\link[=column_width]{column_width()}},
\code{\link[=create_report]{create_report()}},
\code{\link[=dyn_table]{dyn_table()}},
\code{\link[=inbo_website]{inbo_website()}},
\code{\link[=prepare_cover]{prepare_cover()}},
\code{\link[=references]{references()}},
\code{\link[=render_natbib]{render_natbib()}},
\code{\link[=validate_doi]{validate_doi()}}
}
\concept{utils}
