% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_width.R, R/ebook.R, R/handouts.R,
%   R/inbo_rapport_css.R, R/minutes.R, R/mission.R, R/pdf_report.R, R/poster.R,
%   R/slides.R
\name{kolom_breedte}
\alias{kolom_breedte}
\alias{kolom_start}
\alias{ebook}
\alias{inbo_ebook}
\alias{inbo_handouts}
\alias{inbo_rapport_css}
\alias{inbo_verslag}
\alias{inbo_zending}
\alias{inbo_rapport}
\alias{report}
\alias{inbo_poster}
\alias{inbo_slides}
\title{Deprecated functions}
\usage{
kolom_breedte(aantal = 3, marge = 20, witruimte = 10)

kolom_start(aantal = 3, start = 1, marge = 20, witruimte = 10)

ebook()

inbo_ebook()

inbo_handouts(
  toc = TRUE,
  nup = c("8", "1", "1plus", "2", "3", "3plus", "4", "4plus", "6"),
  ...
)

inbo_rapport_css(format = "html")

inbo_verslag(
  present = "",
  absent = "",
  chair = "",
  floatbarrier = c(NA, "section", "subsection", "subsubsection"),
  includes = NULL,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  lang = "dutch",
  keep_tex = FALSE,
  fig_crop = TRUE,
  pandoc_args = NULL,
  ...
)

inbo_zending(
  conference,
  conferencedate,
  conferenceplace,
  website = "",
  reportdate,
  colleagues = "",
  floatbarrier = c(NA, "section", "subsection", "subsubsection"),
  includes = NULL,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  lang = "dutch",
  keep_tex = FALSE,
  fig_crop = TRUE,
  pandoc_args = NULL,
  ...
)

inbo_rapport(fig_crop = "auto", includes = NULL, pandoc_args = NULL, ...)

report(fig_crop = "auto", includes = NULL, pandoc_args = NULL, ...)

inbo_poster(
  subtitle,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  lang = "english",
  email = "info@inbo.be",
  keep_tex = FALSE,
  fig_crop = TRUE,
  includes = NULL,
  pandoc_args = NULL,
  ...
)

inbo_slides(toc = TRUE, ...)
}
\arguments{
\item{aantal}{The desired number of columns with equal width.}

\item{marge}{the horizontal distance between columns in mm.}

\item{witruimte}{the width of the whitespace edge around the poster in mm.}

\item{start}{the number of the column}

\item{toc}{display a table of content after the title slide}

\item{nup}{Number of slides on a page.
Defaults to \code{8}.
The 'plus' layouts leave blank space for recipients to make handwritten notes
next to each slide.
Note that the \code{2}, \verb{3plus}, \code{4} and \code{6} layouts are intended for slides in
the 4:3 aspect ratio, while the \code{3}, \verb{4plus} and \code{8} layouts are intended for
widescreen slides like the 16:10 aspect ratio.}

\item{...}{currently ignored}

\item{format}{defunct}

\item{present}{names of all persons who attended the meeting}

\item{absent}{names of all persons who couldn't make it}

\item{chair}{chair of the meeting}

\item{floatbarrier}{where to place automatic floatbarriers}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{codesize}{relative font size for code}

\item{lang}{main language}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{fig_crop}{Whether to crop PDF figures with the command
\command{pdfcrop}. This requires the tools \command{pdfcrop} and
\command{ghostscript} to be installed. By default, \code{fig_crop = TRUE}
if these two tools are available.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{conference}{the name of the conference or workshop}

\item{conferencedate}{the date of the conference}

\item{conferenceplace}{the location of the conference}

\item{website}{the website of the conference}

\item{reportdate}{the date of this report}

\item{colleagues}{name of other colleagues attending the same conference}

\item{subtitle}{An optional subtitle}

\item{email}{The email address to display at the bottom.
Defaults to info@inbo.be}
}
\description{
\tabular{ll}{
   old function \tab new function \cr
   \code{ìnbo_ebook()} \tab \code{epub_book()} \cr
   \code{ìnbo_handouts()} \tab \code{handouts()} \cr
   \code{ìnbo_poster()} \tab \code{poster()} \cr
   \code{ìnbo_rapport()} \tab \code{pdf_report()} \cr
   \code{inbo_rapport_css()} \tab \code{report_template()} \cr
   \code{ìnbo_slides()} \tab \code{slides()} \cr
   \code{ìnbo_verslag()} \tab \code{minutes()} \cr
   \code{report()} \tab \code{pdf_report()} \cr
}
}
\concept{deprecated}
