% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebook.R
\name{epub_book}
\alias{epub_book}
\title{Style for e-book}
\usage{
epub_book()
}
\description{
A version of \code{\link[bookdown:epub_book]{bookdown::epub_book()}} with a different styling.
}
\section{Available YAML headers options shared among output formats}{

\itemize{
\item \strong{\code{title}}: the main title
\item \code{subtitle}: the optional subtitle
\item \strong{\code{author}}: the authors formatted as indicated in the section below.
\item \code{flandersfont}: if set to \code{TRUE}, use the Flanders Art Sans font.
When not set, or set to \code{FALSE} or empty, use the Calibri font.
\item \code{style}: Defaults to "INBO" when missing.
\itemize{
\item "INBO" for the INBO style.
\item "Vlaanderen" for the generic style of the Flemish Goverment in Dutch.
\item "Flanders" for the generic style of the Flemish Goverment in a language
other than Dutch.
}
\item \code{lang}: the main language of the document as \href{https://datatracker.ietf.org/doc/html/rfc5646}{RFC 5646} tags.
\code{style} "INBO" and "Vlaanderen" require "nl" as \code{lang}.
\code{style} "Flanders" uses "en" as default and can use other languages (but
not "nl").
\itemize{
\item "nl" for Dutch
\item "en" for English
\item "fr" for French
}
\item \code{codesize}: Relative size of R code compared to normal text.
Defaults to "footnotesize".
All options going from large to small are: "normalsize", "small",
"footnotesize", "scriptsize", "tiny".
"normalsize" implies the same size as normal text.
}
}

\section{Formatting author information}{


Quotes are required when a part contains spaces.
If the author doesn't have an ORCID, you can omit that line.
However, note that the INBO policy is that all scientific personnel is
required to obtain an ORCID from https://orcid.org.\preformatted{  - firstname: "Authors first name"
    name: "Authors last name"
    email: "local-part@domain.com"
    orcid: 0000-0002-1825-0097
}
}

\section{Available YAML headers options specific for reports}{


Bold items are required.
Omitting them results in an ugly warning in the rendered document.
This is deliberate done so you can prepare the document without adding the
information.
But the ugly warnings nudges you to add the required information before
publication.
\itemize{
\item \strong{\code{shortauthor}}: abbreviated list of authors.
Required for the citation in the colophon.
Also displayed at the bottom of each page in the web version of the report.
\item \strong{\code{corresponding}}: e-mail of the corresponding author.
Required for the colophon.
\item \strong{\code{reviewer}}: names of the reviewers.
Use the same syntax as for the authors.
\item \strong{\code{year}}: year of publication.
Used in the same places as \code{shortauthor}.
\item \strong{\code{cover_photo}}: the relative path to the image you want on the cover.
\item \strong{\code{cover_description}}: Description of the cover photo, including information
on the author of the picture and license information.
\item \code{cover} the relative path to a pdf with the cover provided by the graphical
designer.
The first page of this pdf will be prepended to the pdf version of the
report.
It becomes the title page in the web and e-book version of the report.
And also shows as a thumbnail in the floating table of content in the web
version.
\item \strong{\code{doi}}: the Digital Object Identifier of the report.
Used in conjunction with \code{shortauthor} and \code{year}.
\item \strong{\code{reportnr}}: Report number assigned by the INBO library.
\item \code{ordernr}: Optional reference number specified by the client.
\item \strong{\code{depotnr}}: Report number assigned by the INBO library.
\item \code{client}: Optionally the name, address and website of the client.
\item \code{client_logo}: Optionally the logo of the client.
Only used when \code{client} is set.
\item \code{cooperation}: Optionally the name, address and website of the organisation
you collaborated with.
\item \code{cooperation_logo}: Optionally the logo of the collaborator.
Only used when \code{cooperation} is set.
\item \code{geraardsbergen}: Set this to \code{TRUE} to add the address of INBO
Geraardsbergen to the colophon.
If not set or set to \code{FALSE} add the address of INBO Brussel.
}
}

\seealso{
Other output: 
\code{\link{gitbook}()},
\code{\link{handouts}()},
\code{\link{minutes}()},
\code{\link{mission}()},
\code{\link{pdf_report}()},
\code{\link{poster}()},
\code{\link{slides}()}
}
\concept{output}
