% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitbook.R
\name{gitbook}
\alias{gitbook}
\title{Output format for a web version of the report}
\usage{
gitbook(code_folding = c("none", "show", "hide"))
}
\arguments{
\item{code_folding}{Enable document readers to toggle the display of R code
chunks. Specify \code{"none"} to display all code chunks. Specify
\code{"hide"} or \code{"show"} to hide or show all R code chunks by
default, and let readers toggle the states on browsers. See the
\emph{Code folding}}
}
\description{
A version of \code{\link[bookdown:gitbook]{bookdown::gitbook()}} with a different styling.
}
\section{Available YAML headers options shared among output formats}{

\itemize{
\item \strong{\code{title}}: the main title
\item \code{subtitle}: the optional subtitle
\item \strong{\code{author}}: the authors formatted as indicated in the section below.
\item \code{flandersfont}: if set to \code{TRUE}, use the Flanders Art Sans font.
When not set, or set to \code{FALSE} or empty, use the Calibri font.
\item \code{style}: Defaults to "INBO" when missing.
\itemize{
\item \code{"INBO"} for the INBO style.
\item \code{"Vlaanderen"} for the generic style of the Flemish Government in
Dutch.
\item \code{"Flanders"} for the generic style of the Flemish Government in a
language other than Dutch.
}
\item \code{lang}: the main language of the document as \href{https://datatracker.ietf.org/doc/html/rfc5646}{RFC 5646} tags.
\code{style} \code{"INBO"} and \code{"Vlaanderen"} require \code{"nl"} as \code{lang}.
\code{style} \code{"Flanders"} uses \code{"en"} as default and can use other languages
(but not \code{"nl"}).
\itemize{
\item \code{"nl"} for Dutch
\item \code{"en"} for English
\item \code{"fr"} for French
}
\item \code{codesize}: Relative size of R code compared to normal text.
Defaults to \code{"footnotesize"}.
All options going from large to small are: \code{"normalsize"}, \code{"small"},
\code{"footnotesize"}, \code{"scriptsize"}, \code{"tiny"}.
\code{"normalsize"} implies the same size as normal text.
}
}

\section{Formatting author information}{


Quotes are required when a part contains spaces.
If the author doesn't have an ORCID, you can omit that line.
However, note that the INBO policy is that all scientific personnel is
required to obtain an ORCID from https://orcid.org.
Add \code{corresponding: true} to the corresponding author.

\if{html}{\out{<div class="sourceCode">}}\preformatted{  - name:
     given: "Authors first name"
     family: "Authors last name"
    email: "local-part@domain.com"
    orcid: 0000-0002-1825-0097
    corresponding: true
}\if{html}{\out{</div>}}
}

\section{Available YAML headers options specific for reports}{


Bold items are required when creating a public report.
Omitting them results in an ugly warning in the rendered document.
This is deliberate done so you can prepare the document without adding the
information.
But the ugly warnings nudges you to add the required information before
publication.
\itemize{
\item \code{public_report}: A logical value indicating whether the report will be
published or not.
Setting \code{public_report: FALSE} will alter the colophon.
Omitting \code{public_report} is equivalent to \code{public_report: TRUE}.
An internal report can't have a DOI.
Setting a DOI is equivalent to \code{public_report: TRUE}.
\item \strong{\code{reviewer}}: names of the reviewers.
Use the same syntax as for the authors.
\item \strong{\code{year}}: year of publication.
Used in the same places as \code{shortauthor}.
\item \strong{\code{cover_photo}}: the relative path to the image you want on the cover.
\item \strong{\code{cover_description}}: Description of the cover photo, including
information on the author of the picture and license information.
\item \code{cover} the relative path to a pdf with the cover provided by the graphical
designer.
The first page of this pdf will be prepended to the pdf version of the
report.
It becomes the title page in the web and e-book version of the report.
And also shows as a thumbnail in the floating table of content in the web
version.
\item \strong{\code{doi}}: the Digital Object Identifier of the report.
Used in conjunction with \code{shortauthor} and \code{year}.
\item \strong{\code{reportnr}}: Report number assigned by the INBO library.
\item \code{ordernr}: Optional reference number specified by the client.
\item \strong{\code{depotnr}}: Report number assigned by the INBO library.
\item \code{client}: Optionally the name, address and website of the client.
\item \code{client_logo}: Optionally the logo of the client.
Only used when \code{client} is set.
\item \code{cooperation}: Optionally the name, address and website of the organisation
you collaborated with.
\item \code{cooperation_logo}: Optionally the logo of the collaborator.
Only used when \code{cooperation} is set.
\item \code{geraardsbergen}: Set this to \code{TRUE} to add the address of INBO
Geraardsbergen to the colophon.
If not set or set to \code{FALSE} add the address of INBO Brussels.
}
}

\section{Available YAML headers options specific for \code{gitbook}}{

\itemize{
\item \code{github-repo}: You can optionally set the \code{github} repo in
\code{organisation/repository} format (e.g. \code{inbo/INBOmd}).
This will add an 'edit' button to the source file on GitHub.
Note that only people with write access can edit the files directly in the
repository.
Others can only edit the source code in a fork of the repository and
suggest the changes through a pull request.
\item \code{split_by}: How to name the HTML output files from the book:
\itemize{
\item \code{chapter+number} means split the file by the first-level headers;
\item \code{section+number} means the second-level headers.
For \code{chapter+number} and \code{section+number}, the HTML file names
will be determined by the header ID's, e.g. the file name for the first
chapter with a chapter title \code{# Introduction} will be
\file{1-introduction.html}.
Note that you need to place at least one level 1 heading in \code{index.Rmd}.
Otherwise you don't get an \code{index.html} file.
}
}
}

\seealso{
Other output: 
\code{\link{epub_book}()},
\code{\link{handouts}()},
\code{\link{minutes}()},
\code{\link{mission}()},
\code{\link{pdf_report}()},
\code{\link{poster}()},
\code{\link{slides}()}
}
\concept{output}
