% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_gitbook.R
\name{inbo_gitbook}
\alias{inbo_gitbook}
\title{INBO version of bookdown::gitbook}
\usage{
inbo_gitbook(
  split_by = c("chapter+number", "section+number"),
  lang = c("nl", "en"),
  style = c("INBO", "Vlaanderen", "Flanders")
)
}
\arguments{
\item{split_by}{How to name the HTML output files from the book:
\code{chapter+number} means split the file by the first-level headers;
\code{section+number} means the second-level headers.
For \code{chapter+number} and \code{section+number}, the HTML filenames will
be determined by the header ID's, e.g. the filename for the first chapter
with a chapter title \code{# Introduction} will be
\file{1-introduction.html}.}

\item{lang}{the language used in the colophon.}

\item{style}{What template to use.
Use \code{"INBO"} for an INBO report in Dutch.
Use \code{"Vlaanderen"} for a report in Dutch written by more than one Flemish
government agency.
Use \code{"Flanders"} for a report in another language.}
}
\description{
INBO version of bookdown::gitbook
}
\seealso{
Other output: 
\code{\link{inbo_handouts}()},
\code{\link{inbo_poster}()},
\code{\link{inbo_rapport}()},
\code{\link{inbo_slides}()},
\code{\link{inbo_verslag}()},
\code{\link{inbo_zending}()}
}
\concept{output}
