% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_handouts.R
\name{inbo_handouts}
\alias{inbo_handouts}
\title{Use the handout version of slides with the INBO theme
The only difference between inbo_slides() and inbo_handouts() is that
inbo_slides() can have progressive slides whereas inbo_handouts() only
displays the final slide.}
\usage{
inbo_handouts(
  subtitle,
  location,
  institute,
  cover,
  cover_offset,
  cover_hoffset,
  cover_horizontal = TRUE,
  toc_name,
  fontsize,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  lang = "dutch",
  slide_level = 2,
  keep_tex = FALSE,
  toc = TRUE,
  website = "www.vlaanderen.be/inbo",
  theme = c("inbo", "vlaanderen"),
  flandersfont = FALSE,
  ...
)
}
\arguments{
\item{keep_tex}{Keep the tex file.
Defaults to \code{FALSE}.}

\item{toc}{display a table of content after the title slide}

\item{...}{currently ignored}
}
\description{
Use the handout version of slides with the INBO theme
The only difference between inbo_slides() and inbo_handouts() is that
inbo_slides() can have progressive slides whereas inbo_handouts() only
displays the final slide.
}
\seealso{
Other output: 
\code{\link{inbo_ebook}()},
\code{\link{inbo_gitbook}()},
\code{\link{inbo_poster}()},
\code{\link{inbo_rapport}()},
\code{\link{inbo_slides}()},
\code{\link{inbo_verslag}()},
\code{\link{inbo_zending}()}
}
\concept{output}
