% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_rapport_2015.R
\name{inbo_rapport_2015}
\alias{inbo_rapport_2015}
\title{Create a report with the INBO theme version 2015}
\usage{
inbo_rapport_2015(subtitle, reportnr, ordernr, floatbarrier = c(NA, "section",
  "subsection", "subsubsection"), natbib, lang = "dutch", keep_tex = FALSE,
  fig_crop = TRUE, ...)
}
\arguments{
\item{subtitle}{An optional subtitle}

\item{reportnr}{The report number}

\item{ordernr}{The order number}

\item{floatbarrier}{Should float barriers be placed? Defaults to NA (no extra float barriers). Options are "section", "subsection" and "subsubsection".}

\item{natbib}{The bibliography file for natbib}

\item{lang}{The language of the document. Defaults to "dutch"}

\item{keep_tex}{Keep the tex file. Defaults to FALSE.}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures}

\item{...}{extra parameters: see details}
}
\description{
Create a report with the INBO theme version 2015
}
\details{
Available extra parameters:
\itemize{
  \item lof: display a list of figures. Defaults tot TRUE
  \item lot: display a list of tables. Defaults tot TRUE
  \item hyphenation: the correct hyphenation for certain words
  \item dankwoord: path to LaTeX file with dankwoord
  \item voorwoord: path to LaTeX file with voorwoord
  \item samenvatting: path to LaTeX file with samenvatting
  \item abstract: path to LaTeX file with english abstract
  \item appendix: path to LaTeX file with appendices
}
}

