% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_rapport_css.R
\name{inbo_rapport_css}
\alias{inbo_rapport_css}
\title{set correct css file and bookdown template}
\usage{
inbo_rapport_css(format = "html")
}
\arguments{
\item{format}{the required output format. Either "html" or "epub"}
}
\description{
set correct css file and bookdown template
}
\seealso{
Other utils: \code{\link{check_dependencies}},
  \code{\link{kolom_breedte}}, \code{\link{kolom_start}},
  \code{\link{render_natbib}}
}
\concept{utils}
