% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inbo_slides}
\alias{inbo_slides}
\title{Use slides with the INBO theme}
\usage{
inbo_slides(subtitle, location, institute, cover, cover_offset, toc_name,
  fontsize, natbib, natbib_title, natbib_style, lang = "dutch",
  slide_level = 2, keep_tex = FALSE)
}
\arguments{
\item{subtitle}{The date and place of the event}

\item{location}{The date and place of the event}

\item{institute}{The aviliation of the authors}

\item{cover}{The filename of the cover image}

\item{cover_offset}{An optional offset for the cover image}

\item{toc_name}{Name of the table of contents. Defaults to "Overzicht".}

\item{fontsize}{The fontsite of the document. Defaults to 10pt.}

\item{natbib}{The bibliography file for natbib}

\item{natbib_title}{The title of the bibliography}

\item{lang}{The language of the document. Defaults to "dutch"}

\item{slide_level}{Indicate which heading level is used for the frame titles}

\item{keep_tex}{Keep the tex file. Defaults to FALSE.}

\item{natbit_style}{The style of the bibliography}
}
\description{
Use slides with the INBO theme
}

