% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolom_breedte.R
\name{kolom_breedte}
\alias{kolom_breedte}
\title{Bereken de breedte van een A0 kolom}
\usage{
kolom_breedte(aantal = 3, marge = 20, witrand = 10)
}
\arguments{
\item{aantal}{het gewenste aantal kolommen van dezelfde breedte}

\item{marge}{de breedte van de afstand tussen de kolommen in mm}

\item{witrand}{de breedte van de rand rond de poster in mm}
}
\description{
Bereken de breedte van een A0 kolom
}
\seealso{
Other utils: 
\code{\link{check_dependencies}()},
\code{\link{cover_info}()},
\code{\link{dyn_table}()},
\code{\link{inbo_rapport_template}()},
\code{\link{kolom_start}()},
\code{\link{references}()},
\code{\link{render_natbib}()}
}
\concept{utils}
