% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolom_breedte.R
\name{kolom_start}
\alias{kolom_start}
\title{Bereken de startpositie van een A0 kolom}
\usage{
kolom_start(aantal = 3, start = 1, marge = 20, witrand = 10)
}
\arguments{
\item{aantal}{het gewenste aantal kolommen van dezelfde breedte}

\item{start}{het nummer van de kolom}

\item{marge}{de breedte van de afstand tussen de kolommen in mm}

\item{witrand}{de breedte van de rand rond de poster in mm}
}
\description{
Bereken de startpositie van een A0 kolom
}
\seealso{
Other utils: 
\code{\link{check_dependencies}()},
\code{\link{cover_info}()},
\code{\link{dyn_table}()},
\code{\link{kolom_breedte}()},
\code{\link{references}()},
\code{\link{render_natbib}()}
}
\concept{utils}
