% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minutes.R
\name{minutes}
\alias{minutes}
\title{Create a minutes of a meeting with the INBO corporate identity}
\usage{
minutes(
  present = "",
  absent = "",
  chair = "",
  floatbarrier = c(NA, "section", "subsection", "subsubsection"),
  includes = NULL,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  lang = "dutch",
  keep_tex = FALSE,
  fig_crop = TRUE,
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{present}{names of all persons who attended the meeting}

\item{absent}{names of all persons who couldn't make it}

\item{chair}{chair of the meeting}

\item{floatbarrier}{where to place automatic floatbarriers}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{codesize}{relative font size for code}

\item{lang}{main language}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{fig_crop}{Whether to crop PDF figures with the command
\command{pdfcrop}. This requires the tools \command{pdfcrop} and
\command{ghostscript} to be installed. By default, \code{fig_crop = TRUE}
if these two tools are available.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{extra parameters: see details}
}
\description{
Create a minutes of a meeting with the INBO corporate identity
}
\details{
Available extra parameters:
\itemize{
\item hyphenation: the correct hyphenation for certain words
}
}
\seealso{
Other output: 
\code{\link{epub_book}()},
\code{\link{gitbook}()},
\code{\link{handouts}()},
\code{\link{mission}()},
\code{\link{pdf_report}()},
\code{\link{poster}()},
\code{\link{slides}()}
}
\concept{output}
