% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mission.R
\name{mission}
\alias{mission}
\title{Create a mission report with the INBO corporate identity}
\usage{
mission(
  conference,
  conferencedate,
  conferenceplace,
  website = "",
  reportdate,
  colleagues = "",
  floatbarrier = c(NA, "section", "subsection", "subsubsection"),
  includes = NULL,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  lang = "dutch",
  keep_tex = FALSE,
  fig_crop = TRUE,
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{conference}{the name of the conference or workshop}

\item{conferencedate}{the date of the conference}

\item{conferenceplace}{the location of the conference}

\item{website}{the website of the conference}

\item{reportdate}{the date of this report}

\item{colleagues}{name of other colleagues attending the same conference}

\item{floatbarrier}{where to place automatic float barriers.}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{codesize}{relative font size for code}

\item{lang}{main language}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF.
Note that this argument does not control whether to keep the auxiliary
files (e.g., \file{.aux}) generated by LaTeX when compiling \file{.tex} to
\file{.pdf}. To keep these files, you may set \code{options(tinytex.clean =
FALSE)}.}

\item{fig_crop}{Whether to crop PDF figures with the command
\command{pdfcrop}. This requires the tools \command{pdfcrop} and
\command{ghostscript} to be installed. By default, \code{fig_crop = TRUE}
if these two tools are available.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{extra parameters: see details}
}
\description{
Create a mission report with the INBO corporate identity
}
\details{
Available extra parameters:
\itemize{
\item hyphenation: the correct hyphenation for certain words
}
}
\seealso{
Other output: 
\code{\link{epub_book}()},
\code{\link{gitbook}()},
\code{\link{handouts}()},
\code{\link{minutes}()},
\code{\link{pdf_report}()},
\code{\link{poster}()},
\code{\link{slides}()}
}
\concept{output}
