% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_report.R
\name{pdf_report}
\alias{pdf_report}
\title{Create a report with the Flemish corporate identity}
\usage{
pdf_report(fig_crop = "auto", includes = NULL, pandoc_args = NULL, ...)
}
\arguments{
\item{fig_crop}{Whether to crop PDF figures with the command
\command{pdfcrop}. This requires the tools \command{pdfcrop} and
\command{ghostscript} to be installed. By default, \code{fig_crop = TRUE}
if these two tools are available.}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{currently ignored}
}
\description{
Create a report with the Flemish corporate identity
}
\section{Available YAML headers options shared among output formats}{

\itemize{
\item \strong{\code{title}}: the main title
\item \code{subtitle}: the optional subtitle
\item \strong{\code{author}}: the authors formatted as indicated in the section below.
\item \code{flandersfont}: if set to \code{TRUE}, use the Flanders Art Sans font.
When not set, or set to \code{FALSE} or empty, use the Calibri font.
\item \code{style}: Defaults to "INBO" when missing.
\itemize{
\item \code{"INBO"} for the INBO style.
\item \code{"Vlaanderen"} for the generic style of the Flemish Government in
Dutch.
\item \code{"Flanders"} for the generic style of the Flemish Government in a
language other than Dutch.
}
\item \code{lang}: the main language of the document as \href{https://datatracker.ietf.org/doc/html/rfc5646}{RFC 5646} tags.
\code{style} \code{"INBO"} and \code{"Vlaanderen"} require \code{"nl"} as \code{lang}.
\code{style} \code{"Flanders"} uses \code{"en"} as default and can use other languages
(but not \code{"nl"}).
\itemize{
\item \code{"nl"} for Dutch
\item \code{"en"} for English
\item \code{"fr"} for French
}
\item \code{codesize}: Relative size of R code compared to normal text.
Defaults to \code{"footnotesize"}.
All options going from large to small are: \code{"normalsize"}, \code{"small"},
\code{"footnotesize"}, \code{"scriptsize"}, \code{"tiny"}.
\code{"normalsize"} implies the same size as normal text.
}
}

\section{Formatting author information}{


Quotes are required when a part contains spaces.
If the author doesn't have an ORCID, you can omit that line.
However, note that the INBO policy is that all scientific personnel is
required to obtain an ORCID from https://orcid.org.
Add \code{corresponding: true} to the corresponding author.

\if{html}{\out{<div class="sourceCode">}}\preformatted{  - name:
     given: "Authors first name"
     family: "Authors last name"
    email: "local-part@domain.com"
    orcid: 0000-0002-1825-0097
    corresponding: true
}\if{html}{\out{</div>}}
}

\section{Available YAML headers options specific for reports}{


Bold items are required when creating a public report.
Omitting them results in an ugly warning in the rendered document.
This is deliberate done so you can prepare the document without adding the
information.
But the ugly warnings nudges you to add the required information before
publication.
\itemize{
\item \code{public_report}: A logical value indicating whether the report will be
published or not.
Setting \code{public_report: FALSE} will alter the colophon.
Omitting \code{public_report} is equivalent to \code{public_report: TRUE}.
An internal report can't have a DOI.
Setting a DOI is equivalent to \code{public_report: TRUE}.
\item \strong{\code{reviewer}}: names of the reviewers.
Use the same syntax as for the authors.
\item \strong{\code{year}}: year of publication.
Used in the same places as \code{shortauthor}.
\item \strong{\code{cover_photo}}: the relative path to the image you want on the cover.
\item \strong{\code{cover_description}}: Description of the cover photo, including
information on the author of the picture and license information.
\item \code{cover} the relative path to a pdf with the cover provided by the graphical
designer.
The first page of this pdf will be prepended to the pdf version of the
report.
It becomes the title page in the web and e-book version of the report.
And also shows as a thumbnail in the floating table of content in the web
version.
\item \strong{\code{doi}}: the Digital Object Identifier of the report.
Used in conjunction with \code{shortauthor} and \code{year}.
\item \strong{\code{reportnr}}: Report number assigned by the INBO library.
\item \code{ordernr}: Optional reference number specified by the client.
\item \strong{\code{depotnr}}: Report number assigned by the INBO library.
\item \code{client}: Optionally the name, address and website of the client.
\item \code{client_logo}: Optionally the logo of the client.
Only used when \code{client} is set.
\item \code{cooperation}: Optionally the name, address and website of the organisation
you collaborated with.
\item \code{cooperation_logo}: Optionally the logo of the collaborator.
Only used when \code{cooperation} is set.
\item \code{geraardsbergen}: Set this to \code{TRUE} to add the address of INBO
Geraardsbergen to the colophon.
If not set or set to \code{FALSE} add the address of INBO Brussels.
}
}

\section{Available YAML headers options specific for reports in pdf format}{


Bold items are required.
Omitting them results in an ugly warning in the rendered document.
This is deliberate done so you can prepare the document without adding the
information.
But the ugly warnings nudges you to add the required information before
publication.
\itemize{
\item \strong{\code{embargo}}: The date at which the report is made public on the INBO
website.
\item \code{print}: section only required in case you need a number of printed copies.
When set it requires the following sub items:
\itemize{
\item \code{copies}: the number of copies
\item \code{motivation}: motivate why you need printed copies as the INBO policy
is to use only digital publications.
\item \code{pages}: the number of pages in the full document.
}
\item \code{tocdepth}: which level headers to display.
\itemize{
\item 0: up to chapters (\verb{#})
\item 1: up to section (\verb{##}) default
\item 2: up to subsection (\verb{###})
}
\item \code{lof}: Adds a list of figures to the pdf after the table of content when
set to \code{TRUE}.
Omits the list of figures when missing or set to \code{FALSE}.
\item \code{lot}: Adds a list of tables to the pdf after the table of content and the
list of figures when set to \code{TRUE}.
Omits the list of tables when missing or set to \code{FALSE}.
\item \code{hyphenation}: An optional list of words with a specified hyphenation
pattern.
E.g. \code{hyphenation: "fortran, ergo-no-mic"} forces the word \code{"fortran"} to
remain as a single word.
"ergonomic" can be either the entire word, split into \code{"ergo"} and
\code{"nomic"} or split into \code{"ergono"} and \code{"mic"}.
\item \code{floatbarrier}: A float barrier forces to place all floating figures and
tables before placing the remainder of the document.
You can place them manually by setting \verb{\\floatbarrier}.
This option allows you to set the automatically before a heading.
Defaults to NA (only float barriers before starting a new chapter \verb{#}).
Options are "section" (before \verb{#} and \verb{##}), "subsection" (before \verb{#}, \verb{##}
and \verb{###}) and "subsubsection" (before \verb{#}, \verb{##}, \verb{###} and \verb{####}).
\item \code{other_lang}: the other languages you want to use in the pdf version.
Available options are \code{"nl"}, \code{"en"} and \code{"fr"}
Defaults to all available languages except the main language.
}
}

\section{Multiple languages in pdf}{


You can define some parts of the text to be in a different language than the
main language (e.g. an abstract in a different language).
This is currently limited to Dutch, English and French.
Use \verb{\\bdutch} before and \verb{\\edutch} after the text in Dutch.
Use \verb{\\benglish} before and \verb{\\eenglish} after the text in English.
Use \verb{\\bfrench} before and \verb{\\efrench} after the text in French.
The styles \code{"INBO"} and \code{"Vlaanderen"} have French and English as optional
languages.
The \code{Flanders} style has by default Dutch and French as optional languages.

Setting the language affects the hyphenation pattern and the names of items
like figures, tables, table of contents, list of figures, list of tables,
references, page numbers, ...
}

\seealso{
Other output: 
\code{\link{epub_book}()},
\code{\link{gitbook}()},
\code{\link{handouts}()},
\code{\link{minutes}()},
\code{\link{mission}()},
\code{\link{poster}()},
\code{\link{slides}()}
}
\concept{output}
