% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poster.R
\name{poster}
\alias{poster}
\title{Create a poster with the INBO theme version 2015}
\usage{
poster(
  subtitle,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  lang = "english",
  email = "info@inbo.be",
  keep_tex = FALSE,
  fig_crop = TRUE,
  includes = NULL,
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{subtitle}{An optional subtitle}

\item{codesize}{relative font size for code}

\item{lang}{The language of the document. Defaults to \code{"english"}.}

\item{email}{The email address to display at the bottom.
Defaults to \code{"info@inbo.be"}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF.
Note that this argument does not control whether to keep the auxiliary
files (e.g., \file{.aux}) generated by LaTeX when compiling \file{.tex} to
\file{.pdf}. To keep these files, you may set \code{options(tinytex.clean =
FALSE)}.}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{pandoc_args}{Additional command line options to pass to \code{pandoc}.}

\item{...}{extra parameters: see details}
}
\description{
Create a poster with the INBO theme version 2015
}
\details{
Available extra parameters:
\itemize{
\item \code{hyphenation}: the correct hyphenation for certain words
\item \code{flandersfont}: Use the Flanders Art Sans font.
Defaults to \code{FALSE}.
Note that this requires the font to be present on the system.
\item ORCID: a list of authors. For each author there must a \code{name} and an
\code{ID}.
The \code{ID} is the author's ORCID ID , see https://orcid.org.
This information will be displayed with QR code at the bottom of the
poster.
\item \code{DOI}: a list of documents.
For each document there must a \code{name} and an \code{ID}.
The \code{ID} is the document's DOI, see https://doi.org.
This information will be displayed with QR code at the bottom of the
poster.
}
}
\seealso{
Other output: 
\code{\link{epub_book}()},
\code{\link{gitbook}()},
\code{\link{handouts}()},
\code{\link{minutes}()},
\code{\link{mission}()},
\code{\link{pdf_report}()},
\code{\link{slides}()}
}
\concept{output}
