% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cover.R
\name{prepare_cover}
\alias{prepare_cover}
\title{Prepare a cover image from a PDF file}
\usage{
prepare_cover(input, path = ".")
}
\arguments{
\item{input}{Link to the PDF file which first becomes the cover.}

\item{path}{the path to the \code{index.Rmd} of the bookdown report.}
}
\value{
Invisible \code{NULL}.
}
\description{
Extract the first page of a PDF file.
Store this page as \code{cover.pdf} at \code{path}.
Also convert it to \code{cover.png}.
}
\seealso{
Other utils: 
\code{\link{add_author}()},
\code{\link{add_report_numbers}()},
\code{\link{check_dependencies}()},
\code{\link{column_start}()},
\code{\link{column_width}()},
\code{\link{cover_info}()},
\code{\link{create_report}()},
\code{\link{dyn_table}()},
\code{\link{inbo_website}()},
\code{\link{references}()},
\code{\link{render_natbib}()},
\code{\link{validate_doi}()}
}
\concept{utils}
