% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_natbib.R
\name{render_natbib}
\alias{render_natbib}
\title{Render a Markdown file with \code{natbib} bibliography}
\usage{
render_natbib(
  file,
  path = ".",
  encoding = "UTF-8",
  engine = c("xelatex", "pdflatex"),
  display = TRUE,
  keep = c("none", "all", "tex"),
  clean = TRUE
)
}
\arguments{
\item{file}{the name of the \code{Rmd} file.}

\item{path}{the path of the \code{Rmd} file.}

\item{encoding}{the encoding of the \code{Rmd} file. Default to 'UTF-8'.}

\item{engine}{the LaTeX engine the compile the document.
Defaults to \code{"xelatex"}.}

\item{display}{open the pdf in a reader. Defaults to TRUE.}

\item{keep}{keep intermediate files after successful compilation.
Defaults to "none".}

\item{clean}{TRUE to clean intermediate files created during rendering of the
R markdown into \code{tex}.}
}
\description{
Render a Markdown file with \code{natbib} bibliography
}
\seealso{
Other utils: 
\code{\link{check_dependencies}()},
\code{\link{column_start}()},
\code{\link{column_width}()},
\code{\link{cover_info}()},
\code{\link{create_report}()},
\code{\link{dyn_table}()},
\code{\link{references}()},
\code{\link{validate_doi}()}
}
\concept{utils}
