% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_template.R
\name{report_template}
\alias{report_template}
\title{set correct css file and bookdown template}
\usage{
report_template(format = c("html", "epub"), lang = c("nl", "en"))
}
\arguments{
\item{format}{the required output format. Either "html" or "epub".}

\item{lang}{the language used in the colophon.}
}
\description{
set correct css file and bookdown template
}
\seealso{
Other utils: 
\code{\link{check_dependencies}()},
\code{\link{cover_info}()},
\code{\link{dyn_table}()},
\code{\link{kolom_breedte}()},
\code{\link{kolom_start}()},
\code{\link{references}()},
\code{\link{render_natbib}()}
}
\concept{utils}
