% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_doi.R
\name{validate_doi}
\alias{validate_doi}
\title{validate a DOI}
\usage{
validate_doi(doi)
}
\arguments{
\item{doi}{a string containing the DOI.}
}
\description{
Checks the format of a DOI.
The format must obey https://www.doi.org/doi_handbook/2_Numbering.html#2.2
The DOI should the minimal version.
Hence no \verb{doi:}, \verb{https://doi.org} or other prefixes.
An example of a minimal version is \code{10.21436/inbor.70809860}.
The part before the forward slash consists of two or three sets of digits
separated by a dot.
E.g. \code{10.21436} or \verb{10.21436.1}.
The part after the forward slash consists either of only digits or of
two sets of any character separated by a dot.
}
\seealso{
Other utils: 
\code{\link{add_author}()},
\code{\link{add_report_numbers}()},
\code{\link{check_dependencies}()},
\code{\link{column_start}()},
\code{\link{column_width}()},
\code{\link{cover_info}()},
\code{\link{create_report}()},
\code{\link{dyn_table}()},
\code{\link{references}()},
\code{\link{render_natbib}()}
}
\concept{utils}
