% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_author.R
\name{add_author}
\alias{add_author}
\title{Add an author to the a report}
\usage{
add_author(path = ".")
}
\arguments{
\item{path}{the path to the \code{index.Rmd} of the bookdown report.}
}
\description{
This function will interactively add the information of an author from the
local authors database.
}
\seealso{
Other utils: 
\code{\link{add_report_numbers}()},
\code{\link{check_dependencies}()},
\code{\link{column_start}()},
\code{\link{column_width}()},
\code{\link{cover_info}()},
\code{\link{create_report}()},
\code{\link{dyn_table}()},
\code{\link{references}()},
\code{\link{render_natbib}()},
\code{\link{validate_doi}()}
}
\concept{utils}
