% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{check_dependencies}
\alias{check_dependencies}
\title{Check if the required dependencies of INBOmd are installed
Missing dependencies will be installed automatically}
\usage{
check_dependencies()
}
\description{
Check if the required dependencies of INBOmd are installed
Missing dependencies will be installed automatically
}
\seealso{
Other utils: 
\code{\link{add_report_numbers}()},
\code{\link{column_start}()},
\code{\link{column_width}()},
\code{\link{cover_info}()},
\code{\link{create_report}()},
\code{\link{dyn_table}()},
\code{\link{references}()},
\code{\link{render_natbib}()},
\code{\link{validate_doi}()}
}
\concept{utils}
