% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Create a template for an INBOmd bookdown report}
\usage{
create_report(path = ".", shortname)
}
\arguments{
\item{path}{The folder in which to create the folder containing the report.
Defaults to the current working directory.}

\item{shortname}{The name of the report project.
The location of the folder \code{shortname} depends on the content of \code{path}.
When \code{path} is a subfolder of a git repository, it is changed to the root
of that git repository.
When \code{path} is a \code{checklist::checklist} project, you will find the new report
at \code{path/source/shortname}.
When \code{path} is a \code{checklist::checklist} package, you will find the new report
at \code{path/inst/shortname}.
Otherwise you will find the new report at \code{path/shortname}.}
}
\description{
Create a template for an INBOmd bookdown report
}
\seealso{
Other utils: 
\code{\link{add_author}()},
\code{\link{add_report_numbers}()},
\code{\link{check_dependencies}()},
\code{\link{column_start}()},
\code{\link{column_width}()},
\code{\link{cover_info}()},
\code{\link{dyn_table}()},
\code{\link{references}()},
\code{\link{render_natbib}()},
\code{\link{validate_doi}()}
}
\concept{utils}
