% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_table.R
\name{dyn_table}
\alias{dyn_table}
\title{Use a different table style depending on the output format}
\usage{
dyn_table(x, caption = NULL, rownames = FALSE, escape = TRUE, align, ...)
}
\arguments{
\item{x}{the dataframe}

\item{caption}{the table caption; a character vector or a tag object
generated from \code{htmltools::tags$caption()}}

\item{rownames}{\code{TRUE} (show row names) or \code{FALSE} (hide row names)
or a character vector of row names; by default, the row names are displayed
in the first column of the table if exist (not \code{NULL})}

\item{escape}{whether to escape HTML entities in the table: \code{TRUE} means
to escape the whole table, and \code{FALSE} means not to escape it;
alternatively, you can specify numeric column indices or column names to
indicate which columns to escape, e.g. \code{1:5} (the first 5 columns),
\code{c(1, 3, 4)}, or \code{c(-1, -3)} (all columns except the first and
third), or \code{c('Species', 'Sepal.Length')}; since the row names take
the first column to display, you should add the numeric column indices
by one when using \code{rownames}}

\item{align}{Column alignment: a character vector consisting of \code{'l'}
(left), \code{'c'} (center) and/or \code{'r'} (right). By default or if
\code{align = NULL}, numeric columns are right-aligned, and other columns
are left-aligned. If \code{length(align) == 1L}, the string will be
expanded to a vector of individual letters, e.g. \code{'clc'} becomes
\code{c('c', 'l', 'c')}, unless the output format is LaTeX.}

\item{...}{Other arguments (see Examples and References).}
}
\description{
\itemize{
\item \strong{interactive session}: \code{\link[DT]{datatable}}
\item \strong{html}: \code{\link[DT]{datatable}}
\item \strong{other}: \code{\link[knitr]{kable}}
}
}
