% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_ebook.R
\name{inbo_ebook}
\alias{inbo_ebook}
\title{INBO version of bookdown::epub_book}
\usage{
inbo_ebook(lang = c("nl", "en"), style = c("INBO", "Vlaanderen", "Flanders"))
}
\arguments{
\item{lang}{the language used in the colophon.}
}
\description{
INBO version of bookdown::epub_book
}
\seealso{
Other output: 
\code{\link{inbo_gitbook}()},
\code{\link{inbo_handouts}()},
\code{\link{inbo_poster}()},
\code{\link{inbo_rapport}()},
\code{\link{inbo_slides}()},
\code{\link{inbo_verslag}()},
\code{\link{inbo_zending}()}
}
\concept{output}
