% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inbo_rapport}
\alias{inbo_rapport}
\title{Create a report with the INBO theme}
\usage{
inbo_rapport(subtitle, reportnr, ordernr, year, email, shortauthor,
  office = c("Anderlecht", "Geraardsbergen"), client, storagenr, cover,
  cover_offset, cover_text, natbib, floatbarrier = c(NA, "section",
  "subsection", "subsubsection"), lang = "dutch", keep_tex = FALSE, ...)
}
\arguments{
\item{subtitle}{An optional subtitle}

\item{reportnr}{The report number}

\item{ordernr}{The order number}

\item{year}{The year of publication}

\item{email}{The email of the corresponding author}

\item{shortauthor}{The abbreviation of the authors}

\item{office}{The office of the main author. Defaults to "Anderlecht"}

\item{client}{The name of the client}

\item{storagenr}{The storage number of the report}

\item{cover}{The filename of the cover image}

\item{cover_offset}{An optional offset for the cover image}

\item{cover_text}{A description of the cover image}

\item{natbib}{The bibliography file for natbib}

\item{floatbarrier}{Should float barriers be placed? Defaults to NA (no extra float barriers). Options are "section", "subsection" and "subsubsection".}

\item{lang}{The language of the document. Defaults to "dutch"}

\item{keep_tex}{Keep the tex file. Defaults to FALSE.}

\item{...}{extra parameters}
}
\description{
Create a report with the INBO theme
}

