% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_rapport.R
\name{inbo_rapport}
\alias{inbo_rapport}
\title{Create a report with the Flemish corporate identity}
\usage{
inbo_rapport(
  subtitle,
  reportnr,
  ordernr,
  floatbarrier = c(NA, "section", "subsection", "subsubsection"),
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  style = c("INBO", "Vlaanderen", "Flanders"),
  main_language = "english",
  other_languages = c("french", "dutch"),
  keep_tex = FALSE,
  fig_crop = TRUE,
  includes = NULL,
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{subtitle}{An optional subtitle.}

\item{reportnr}{The report number.
Defaults to the date and time of compilation.}

\item{ordernr}{The optional order number.}

\item{floatbarrier}{Should float barriers be placed?
Defaults to NA (only float barriers before starting a new chapter \verb{#}).
Options are "section" (\verb{##}), "subsection" (\verb{###}) and
"subsubsection" (\verb{####}).}

\item{codesize}{The fontsize of the code, relative to the fontsize of the
text (= normal size).
Allowed values are \code{"normalsize"}, \code{"small"}, \code{"footnotesize"},
\code{"scriptsize"} and \code{"tiny"}.
Defaults to \code{"footnotesize"}.}

\item{style}{What template to use.
Use "INBO" for an INBO report in Dutch.
Use "Vlaanderen" for a report in Dutch written by more than one Flemish
government agency.
Use "Flanders" for a report in English.}

\item{main_language}{The main language of the document.
Only used with \code{style = "Flanders"}.
Defaults to \code{"english"}.
Note that \code{style = "INBO"} and \code{style = "Vlaanderen"} will always use
\code{"dutch"} as the main language.}

\item{other_languages}{A vector of other languages you want to use within the
document.
Only used with \code{style = "Flanders"}.
Defaults to \code{c("dutch", "french")}.
Allowed languages are \code{"dutch"}, \code{"french"} and \code{"english"}.
Note that \code{style = "INBO"} and \code{style = "Vlaanderen"} will always use
\code{c("english", "french")} as the other languages.
Use \verb{\\bdutch} before and \verb{\\edutch} after the text in Dutch.
Use \verb{\\benglish} before and \verb{\\eenglish} after the text in English.
Use \verb{\\bfrench} before and \verb{\\efrench} after the text in French.
You don't need to set this for the main language.}

\item{keep_tex}{Keep the tex file.
Defaults to \code{FALSE}.}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures.}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{extra parameters: see details}
}
\description{
Create a report with the Flemish corporate identity
}
\details{
Available extra parameters:
\itemize{
\item \code{lof}: display a list of figures. Defaults to TRUE
\item \code{lot}: display a list of tables. Defaults to TRUE
\item \code{tocdepth}: which level headers to display.
\itemize{
\item 0: upto chapters (\verb{#})
\item 1: upto section (\verb{##})
\item 2: upto subsection (\verb{###})
\item 3: upto subsubsection (\verb{####}) default
}
\item \code{hyphenation}: the correct hyphenation for certain words.
\item \code{cover}: an optional pdf file.
The first two pages will be prepended to the report.
\item \code{ref_title}: the title of the reference section.
Only used when \code{style = "Flanders"}.
}
}
\seealso{
Other output: 
\code{\link{inbo_handouts}()},
\code{\link{inbo_poster}()},
\code{\link{inbo_slides}()},
\code{\link{inbo_verslag}()},
\code{\link{inbo_zending}()}
}
\concept{output}
