% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_rapport_2015.R
\name{inbo_rapport_2015}
\alias{inbo_rapport_2015}
\title{Create a report with the INBO theme version 2015}
\usage{
inbo_rapport_2015(subtitle, reportnr, ordernr, floatbarrier = c(NA, "section",
  "subsection", "subsubsection"), natbib, lang = "dutch", keep_tex = FALSE,
  ...)
}
\arguments{
\item{subtitle}{An optional subtitle}

\item{reportnr}{The report number}

\item{ordernr}{The order number}

\item{floatbarrier}{Should float barriers be placed? Defaults to NA (no extra float barriers). Options are "section", "subsection" and "subsubsection".}

\item{natbib}{The bibliography file for natbib}

\item{lang}{The language of the document. Defaults to "dutch"}

\item{keep_tex}{Keep the tex file. Defaults to FALSE.}

\item{...}{extra parameters: see details

Available extra parameters:
- hyphenation: the correct hyphenation for certain words
- dankwoord: path to LaTeX file with dankwoord
- voorwoord: path to LaTeX file with voorwoord
- samenvatting: path to LaTeX file with samenvatting
- abstract: path to LaTeX file with english abstract
- appendix: path to LaTeX file with appendices}
}
\description{
Create a report with the INBO theme version 2015
}

