% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_rapport_template.R
\name{inbo_rapport_css}
\alias{inbo_rapport_css}
\title{Deprecated}
\usage{
inbo_rapport_css(format = "html")
}
\arguments{
\item{format}{the required output format. Either "html" or "epub".}
}
\description{
Use \code{inbo_rapport_template()} instead.
}
