% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_slides.R
\name{inbo_slides}
\alias{inbo_slides}
\title{Use slides with the INBO theme}
\usage{
inbo_slides(subtitle, location, institute, cover, cover_offset, cover_hoffset,
  cover_horizontal = TRUE, toc_name, fontsize, codesize = c("footnotesize",
  "scriptsize", "tiny", "small", "normalsize"), natbib, natbib_title,
  natbib_style, lang = "dutch", slide_level = 2, keep_tex = FALSE,
  toc = TRUE, ...)
}
\arguments{
\item{subtitle}{The date and place of the event}

\item{location}{The date and place of the event}

\item{institute}{The aviliation of the authors}

\item{cover}{The filename of the cover image}

\item{cover_offset}{An optional vertical offset for the cover image}

\item{cover_hoffset}{An optional horizontal offset for the cover image}

\item{cover_horizontal}{Scale the cover horizontal (TRUE) or vertical (FALSE)}

\item{toc_name}{Name of the table of contents. Defaults to "Overzicht".}

\item{fontsize}{The fontsite of the document. Defaults to 10pt.}

\item{codesize}{The fontsize of the code, relative to the fontsize of the text (= normal size). Allowed values are "normalsize", "small", "footnotesize", "scriptsize" and  "tiny". Defaults to "footnotesize".}

\item{natbib}{The bibliography file for natbib}

\item{natbib_title}{The title of the bibliography}

\item{natbib_style}{The style of the bibliography}

\item{lang}{The language of the document. Defaults to "dutch"}

\item{slide_level}{Indicate which heading level is used for the frame titles}

\item{keep_tex}{Keep the tex file. Defaults to FALSE.}

\item{toc}{display a table of content after the title slide}

\item{...}{extra parameters}
}
\description{
Use slides with the INBO theme
}

