% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_slides.R
\name{inbo_slides}
\alias{inbo_slides}
\title{Use slides with the INBO theme}
\usage{
inbo_slides(
  subtitle,
  location,
  institute,
  cover,
  cover_offset,
  cover_hoffset,
  cover_horizontal = TRUE,
  slide_logo,
  toc_name,
  fontsize,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  citation_package = c("natbib", "none"),
  natbib_title,
  lang = "dutch",
  slide_level = 2,
  keep_tex = FALSE,
  toc = TRUE,
  website = "www.INBO.be",
  theme = c("inbo", "vlaanderen", "inboenglish"),
  flandersfont = FALSE,
  ...
)
}
\arguments{
\item{subtitle}{The subtitle.}

\item{location}{The date and place of the event.}

\item{institute}{The affiliation of the authors.}

\item{cover}{The filename of the cover image.}

\item{cover_offset}{An optional vertical offset for the cover image.}

\item{cover_hoffset}{An optional horizontal offset for the cover image.}

\item{cover_horizontal}{Scale the cover horizontal (\code{TRUE}) or vertical
(\code{FALSE}).}

\item{slide_logo}{the path to an optional logo displayed on each slide}

\item{toc_name}{Name of the table of contents. Defaults to "Overzicht".}

\item{fontsize}{The fontsite of the document. Defaults to 10pt.}

\item{codesize}{The fontsize of the code, relative to the fontsize of the
text (= normal size).
Allowed values are "normalsize", "small", "footnotesize", "scriptsize" and
"tiny". Defaults to "footnotesize".}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{none} if neither package is to be used.}

\item{natbib_title}{The title of the bibliography}

\item{lang}{The language of the document. Defaults to "dutch"}

\item{slide_level}{Indicate which heading level is used for the frame titles}

\item{keep_tex}{Keep the tex file. Defaults to FALSE.}

\item{toc}{display a table of content after the title slide}

\item{website}{An optional URL to display on the left sidebar.
Defaults to www.INBO.be.}

\item{theme}{The theme to use. Available options are "inbo" and "vlaanderen".}

\item{flandersfont}{If TRUE use the Flanders Art font. If FALSE use Calibri.
Defaults to FALSE.}

\item{...}{extra parameters}
}
\description{
Use slides with the INBO theme
}
\seealso{
Other output: 
\code{\link{inbo_handouts}()},
\code{\link{inbo_poster}()},
\code{\link{inbo_rapport}()},
\code{\link{inbo_verslag}()},
\code{\link{inbo_zending}()}
}
\concept{output}
