% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_verslag.R
\name{inbo_verslag}
\alias{inbo_verslag}
\title{Create a meeting report with the INBO corporate identity}
\usage{
inbo_verslag(
  present = "",
  absent = "",
  chair = "",
  floatbarrier = c(NA, "section", "subsection", "subsubsection"),
  includes = NULL,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  lang = "dutch",
  keep_tex = FALSE,
  fig_crop = TRUE,
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{present}{names of all persons who attended the meeting}

\item{absent}{names of all persons who couldn't make it}

\item{chair}{chair of the meeting}

\item{floatbarrier}{Should float barriers be placed?
Defaults to NA (only float barriers before starting a new chapter \verb{#}).
Options are "section" (\verb{##}), "subsection" (\verb{###}) and
"subsubsection" (\verb{####}).}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{extra parameters: see details}
}
\description{
Create a meeting report with the INBO corporate identity
}
\details{
Available extra parameters:
\itemize{
\item hyphenation: the correct hyphenation for certain words
}
}
\seealso{
Other output: 
\code{\link{inbo_ebook}()},
\code{\link{inbo_gitbook}()},
\code{\link{inbo_handouts}()},
\code{\link{inbo_poster}()},
\code{\link{inbo_rapport}()},
\code{\link{inbo_slides}()},
\code{\link{inbo_zending}()}
}
\concept{output}
