% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_verslag.R
\name{inbo_verslag}
\alias{inbo_verslag}
\title{Create a meeting report with the INBO corporate identity}
\usage{
inbo_verslag(present = "", absent = "", chair = "",
  floatbarrier = c(NA, "section", "subsection", "subsubsection"),
  citation_package = c("natbib", "none"), includes = NULL,
  codesize = c("footnotesize", "scriptsize", "tiny", "small",
  "normalsize"), lang = "dutch", keep_tex = FALSE, fig_crop = TRUE,
  pandoc_args = NULL, ...)
}
\arguments{
\item{present}{names of all persons who attended the meeting}

\item{absent}{names of all persons who couldn't make it}

\item{chair}{chair of the meeting}

\item{floatbarrier}{Should float barriers be placed? Defaults to NA (no extra float barriers). Options are "section", "subsection" and "subsubsection".}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{none} if neither package is to be used.}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}

\item{codesize}{The fontsize of the code, relative to the fontsize of the text (= normal size). Allowed values are "normalsize", "small", "footnotesize", "scriptsize" and  "tiny". Defaults to "footnotesize".}

\item{lang}{The language of the document. Defaults to "dutch"}

\item{keep_tex}{Keep the tex file. Defaults to FALSE.}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{extra parameters: see details}
}
\description{
Create a meeting report with the INBO corporate identity
}
\details{
Available extra parameters:
\itemize{
\item hyphenation: the correct hyphenation for certain words
}
}
