% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_website.R
\name{inbo_website}
\alias{inbo_website}
\title{Render the report into a zip file for the INBO website}
\usage{
inbo_website(path = ".")
}
\arguments{
\item{path}{The path to the directory where the report is located}
}
\description{
Render the report into a zip file for the INBO website
}
\seealso{
Other utils:
\code{\link[=add_author]{add_author()}},
\code{\link[=add_report_numbers]{add_report_numbers()}},
\code{\link[=check_dependencies]{check_dependencies()}},
\code{\link[=column_start]{column_start()}},
\code{\link[=column_width]{column_width()}},
\code{\link[=cover_info]{cover_info()}},
\code{\link[=create_report]{create_report()}},
\code{\link[=dyn_table]{dyn_table()}},
\code{\link[=prepare_cover]{prepare_cover()}},
\code{\link[=references]{references()}},
\code{\link[=render_natbib]{render_natbib()}},
\code{\link[=validate_doi]{validate_doi()}}
}
\concept{utils}
