% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_zending.R
\name{inbo_zending}
\alias{inbo_zending}
\title{Create a mission report with the INBO corporate identity}
\usage{
inbo_zending(
  conference,
  conferencedate,
  conferenceplace,
  website = "",
  reportdate,
  colleagues = "",
  floatbarrier = c(NA, "section", "subsection", "subsubsection"),
  includes = NULL,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  lang = "dutch",
  keep_tex = FALSE,
  fig_crop = TRUE,
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{conference}{the name of the conference or workshop}

\item{conferencedate}{the date of the conference}

\item{conferenceplace}{the location of the conference}

\item{website}{the website of the conference}

\item{reportdate}{the date of this report}

\item{colleagues}{name of other colleagues attending the same conference}

\item{floatbarrier}{where to place automatic floatbarriers}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{codesize}{relative font size for code}

\item{lang}{main language}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{extra parameters: see details}
}
\description{
Create a mission report with the INBO corporate identity
}
\details{
Available extra parameters:
\itemize{
\item hyphenation: the correct hyphenation for certain words
}
}
\seealso{
Other output: 
\code{\link{inbo_ebook}()},
\code{\link{inbo_gitbook}()},
\code{\link{inbo_handouts}()},
\code{\link{inbo_poster}()},
\code{\link{inbo_rapport}()},
\code{\link{inbo_slides}()},
\code{\link{inbo_verslag}()}
}
\concept{output}
