% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbo_zending.R
\name{inbo_zending}
\alias{inbo_zending}
\title{Create a mission report with the INBO corporate identity}
\usage{
inbo_zending(conference, conferencedate, conferenceplace, website = "",
  reportdate, colleagues = "", floatbarrier = c(NA, "section",
  "subsection", "subsubsection"), citation_package = c("natbib", "none"),
  includes = NULL, codesize = c("footnotesize", "scriptsize", "tiny",
  "small", "normalsize"), lang = "dutch", keep_tex = FALSE,
  fig_crop = TRUE, pandoc_args = NULL, ...)
}
\arguments{
\item{conference}{the name of the conference or workshop}

\item{conferencedate}{the date of the conference}

\item{conferenceplace}{the location of the conference}

\item{website}{the website of the conference}

\item{reportdate}{the date of this report}

\item{colleagues}{name of other colleagues attending the same conference}

\item{floatbarrier}{Should float barriers be placed? Defaults to NA (no extra float barriers). Options are "section", "subsection" and "subsubsection".}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{none} if neither package is to be used.}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}

\item{codesize}{The fontsize of the code, relative to the fontsize of the text (= normal size). Allowed values are "normalsize", "small", "footnotesize", "scriptsize" and  "tiny". Defaults to "footnotesize".}

\item{lang}{The language of the document. Defaults to "dutch"}

\item{keep_tex}{Keep the tex file. Defaults to FALSE.}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{extra parameters: see details}
}
\description{
Create a mission report with the INBO corporate identity
}
\details{
Available extra parameters:
\itemize{
\item hyphenation: the correct hyphenation for certain words
}
}
