% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poster.R
\name{poster}
\alias{poster}
\title{Create a poster with the INBO theme version 2015}
\usage{
poster(
  subtitle,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  lang = "english",
  email = "info@inbo.be",
  keep_tex = FALSE,
  fig_crop = TRUE,
  includes = NULL,
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{subtitle}{An optional subtitle}

\item{codesize}{relative font size for code}

\item{lang}{The language of the document. Defaults to "english".}

\item{email}{The email address to display at the bottom.
Defaults to info@inbo.be}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{extra parameters: see details}
}
\description{
Create a poster with the INBO theme version 2015
}
\details{
Available extra parameters:
\itemize{
\item hyphenation: the correct hyphenation for certain words
\item flandersfont: Use the Flanders Art Sans font. Defaults to FALSE.
Note that this requires the font to be present on the system.
\item ORCID: a list of authors. For each author there must a \code{name} and an
\code{ID}. The \code{ID} is the author's ORCID ID , ee https://orcid.org.
This information will be displayed with QR code at the bottom of the
poster.
\item DOI: a list of documents. For each documentr there must a \code{name} and an
\code{ID}. The \code{ID} is the documents's DOI, see https://doi.org.
This information will be displayed with QR code at the bottom of the
poster.
}
}
\seealso{
Other output: 
\code{\link{ebook}()},
\code{\link{gitbook}()},
\code{\link{handouts}()},
\code{\link{minutes}()},
\code{\link{mission}()},
\code{\link{pdf_report}()},
\code{\link{slides}()}
}
\concept{output}
