% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{references}
\alias{references}
\title{Insert references at this position}
\usage{
references(appendix = FALSE, part = FALSE)
}
\arguments{
\item{appendix}{Chapters after the references become the appendix.}

\item{part}{Include the appendix as a part (level 0) heading.
Only relevant for pdf output.}
}
\description{
Insert references at this position
}
\seealso{
Other utils: 
\code{\link{add_report_numbers}()},
\code{\link{check_dependencies}()},
\code{\link{column_start}()},
\code{\link{column_width}()},
\code{\link{cover_info}()},
\code{\link{create_report}()},
\code{\link{dyn_table}()},
\code{\link{render_natbib}()},
\code{\link{validate_doi}()}
}
\concept{utils}
