% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsos_article.R
\name{rsos_article}
\alias{rsos_article}
\title{Royal Society Open Science journal format.}
\usage{
rsos_article(..., keep_tex = TRUE, citation_package = "natbib")
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{none} if neither package is to be used.}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating submissions to Royal Society Open Science journals.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "rsos_article", package = "rticles")
}

}
